rename word->mcb
